third_party/WebKit/LayoutTests/xr/getDevicePose_oneframeupdate.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../external/wpt/resources/chromium/webxr-test.js"></script>
<script src="../xr/resources/xr-test-utils.js"></script>
<script src="../xr/resources/test-constants.js"></script>
<canvas id="webgl-canvas"></canvas>

<script>

let testName = "XRFrame getDevicePose updates on the next frame";

let fakeDeviceInitParams = { supportsImmersive: true };

let requestSessionOptions =  [
  { immersive: true },
  { outputContext: getOutputContext() }
];

let testFunction = function(session, t, fakeDeviceController) {
  // Session must have a baseLayer or else frame requests will be ignored.
  session.baseLayer = new XRWebGLLayer(session, gl);

  return session.requestFrameOfReference("eye-level")
    .then((frameOfRef) => new Promise((resolve, reject) => {
      let counter = 0;
      function onFrame(time, vrFrame) {
        session.requestAnimationFrame(onFrame);
        if (counter == 0) {
          t.step( () => {
            // Expecting to not get a pose since none has been supplied
            assert_equals(vrFrame.getDevicePose(frameOfRef), null);

            // Need to have a valid pose or input event's don't process.
            fakeDeviceController.setXRPresentationFrameData(
              VALID_POSE_MATRIX, [{
                eye:"left",
                projectionMatrix: VALID_PROJECTION_MATRIX,
                viewMatrix: VALID_VIEW_MATRIX
              }, {
                eye:"right",
                projectionMatrix: VALID_PROJECTION_MATRIX,
                viewMatrix: VALID_VIEW_MATRIX
              }]);

            // Check that pose does not update pose within the same frame.
            assert_equals(vrFrame.getDevicePose(frameOfRef), null);
          });
        } else {
          t.step( () => {
            // Check that pose was updated.
            let pose = vrFrame.getDevicePose(frameOfRef);
            assert_not_equals(pose, null);

            let poseMatrix = pose.poseModelMatrix;
            assert_not_equals(poseMatrix, null);

            assert_matrices_approx_equal(poseMatrix, VALID_POSE_MATRIX);
          });

          // Finished.
          resolve();
        }
        counter++;
      }

      session.requestAnimationFrame(onFrame);
    }));
};

xr_session_promise_test(
  testFunction, fakeDeviceInitParams, requestSessionOptions, testName);

</script>
